Add TH Quasiquoter for TLA+ expressions
Parser tests/bugs now break the package build because
  [tla| ...]
runs the parser as part of the build of the test executable.

To evaluate the TLA+ test expression, execute:
  $ example-simple-expr

Note the new tlaplus-qq package so users who only want to parse and pretty print TLA+, but not use the Quasiquoter do not need to enable template-haskell (TH). This is the root cause of the orphan instance warnings, but I found it better than auto deriving Lift in the tlaplus Syntax module itself.